Fixed folder structure and added proper adapter and presenter for activity
