+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/LevelUpQuest.iml" filepath="$PROJECT_DIR$/LevelUpQuest.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/td4m3/td4m3.iml" filepath="$PROJECT_DIR$/td4m3/td4m3.iml" />
     </modules>
