--- td4m3/src/main/java/ca/td/td4m3/dashboard/creditCard/contract/MortgageView.kt
@@ -1,4 +1,0 @@
-package ca.td.td4m3.dashboard.creditCard.contract
-
-interface MortgageView {
-}
