+++ td4m3/src/main/java/ca/td/td4m3/dashboard/mortgage/contract/MortgageView.kt
@@ -1,0 +1,5 @@
+package ca.td.td4m3.dashboard.mortgage.contract
+
+interface MortgageView {
+    fun launchMortgageActivity()
+}
