+++ td4m3/src/main/java/ca/td/td4m3/mortgage/adapter/MortgageActivityAdapter.kt
@@ -1,0 +1,9 @@
+package ca.td.td4m3.mortgage.adapter
+
+import com.ngam.rvabstractions.AbstractDataBindAdapter
+
+class MortgageActivityAdapter: AbstractDataBindAdapter() {
+    override fun buildRows() {
+        // TODO
+    }
+}
