+++ td4m3/src/main/java/ca/td/td4m3/mortgage/presenter/MortgageActivityPresenter.kt
@@ -1,0 +1,6 @@
+package ca.td.td4m3.mortgage.presenter
+
+import com.ngam.rvabstractions.AbstractPresenter
+
+class MortgageActivityPresenter: AbstractPresenter() {
+}
