+++ td4m3/src/main/java/ca/td/td4m3/mortgage/view/MortgageActivityView.kt
@@ -1,0 +1,4 @@
+package ca.td.td4m3.mortgage.view
+
+interface MortgageActivityView {
+}
