+++ td4m3/src/main/res/layout/dashboard_layout.xml
@@ -10,7 +10,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/card_top_spacing" />
 
-    <ca.td.td4m3.dashboard.creditCard.cardView.MortgageCardView
+    <ca.td.td4m3.dashboard.mortgage.cardView.MortgageCardView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/card_divider_spacing" />
